break apart api_testrail - p1
